Add lower spirit thresholds and start 2s tick timer when drinking
Added lower spirit thresholds to get rid of some of the issues with Blessing of Wisdom and other similar buffs. Blessing of wisdom has its own tick and was restarting the 2s tick timer. Also made the 5s tick timer end early if the player is drinking; this is critical for players who want to use the timer to 'tick drink' like myself. Tick drinking is running and starting to drink right before the tick, so you can move and drink at the same time.