Simplify how we ticks are tracked.
Instead of tracking the amount of mana we expect to be gained on each tick depending on stats, buffs, and spell usages, we rely on the strict timing of the UNIT_POWER_FREQUENT event. This has the benefit of being more simple and more robust to gear sets with stacked MP5 as well as using consumables restoring mana.

Everything related to parsing buffs and gear is removed, including localization thereof.